Moved PluginConfig to plugin.py as extra file uneccesary, tested & works
